+++ src/app/pages/about-page/about-page-component.ts
@@ -5,7 +5,17 @@ import { Component } from '@angular/core';
   imports: [],
   template: `
     <div class="container-fluid p-3">
-      <ng-content></ng-content>
+      <p>
+        Diese Seite hat keine Funktionalität, sondern dient lediglich als
+        Beispiel für das Navigieren über den Header. Nachfolgend der Demtext,
+        damit es nicht so leer aussieht ;-)
+        <br />
+        <br />
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Vitae, rerum
+        possimus inventore ratione deleniti ea aut accusantium esse quis quaerat
+        dolorem quidem corporis nostrum molestias? Assumenda architecto quidem
+        voluptates officiis!
+      </p>
     </div>
   `,
   styles: ``,
