+++ src/app/pages/tabelle-page/components/filtertabelle/sanitizer/sanitize-result.ts
