+++ src/app/service/unternehmen/unternehmen.model.ts
@@ -1,0 +1,8 @@
+export interface UnternehmenModel {
+  name: string;
+  land: string;
+  countryCode: string;
+  branche: string;
+  co2Verbrauch: number;
+  id: number;
+}
